+++ Plugins/UI/Source/NexusUI/Private/NCanvasUtils.cpp
@@ -1,13 +1,23 @@
-﻿#include "NCanvasUtils.h"
+﻿// Copyright dotBunny Inc. All Rights Reserved.
+// See the LICENSE file at the repository root for more information.
+
+#include "NCanvasUtils.h"
 
 #include "CanvasItem.h"
 #include "CanvasTypes.h"
 
 void FNCanvasUtils::DrawCanvasTextBox(FNMultiLineTextBoxCanvasItem* TextBox, FCanvas* Canvas, FVector2D TopLeftPosition)
 {
+	// Ensure processed
 	if (TextBox->IsDirty())
 	{
 		TextBox->Process(10.f);
+	}
+	
+	// Check we're still good
+	if (!TextBox->HasContent())
+	{
+		return;
 	}
 	
 	FVector2D Padding(10.0f, 10.0f); // X and Y internal padding